fix: PHPStan unused property and PSR-4 autoloading for Git ability classes
- Remove unused $settings constructor injection from RestController (PHPStan:
  'property is never read, only written')
- Extract GitSnapshotAbility, GitDiffAbility, GitRestoreAbility, GitListAbility,
  GitPackageSummaryAbility, and GitRevertPackageAbility from GitAbilities.php
  into separate PSR-4 files so the composer autoloader can resolve them by
  class name without requiring GitAbilities.php to be loaded first
- GitAbilities.php retains only the registration class (no class definitions)

Verified: phpstan analyse → No errors; phpcs → 0 violations; php -l clean
